Split api messages classes into their own file
